mgr/dashboard: "Any" initiator entry appears in the Initiator table after selecting "All Hosts Allowed" during subsystem creation
Fixes: https://tracker.ceph.com/issues/77467

(cherry picked from commit d83f5dcdb905bd7b1de0ffaba951c7b6d491f467)

 Conflicts:
	src/pybind/mgr/dashboard/frontend/src/app/ceph/block/nvmeof-gateway-node/nvmeof-gateway-node.component.ts
- Conflict in nvmeof-gateway-node.component.ts during cherry-pick onto tentacle.
Kept the enum relocation import: NvmeofGatewayNodeMode from ~/app/shared/models/nvmeof.
Kept target-branch (tentacle) service-spec usage: CephServiceSpec only
