--- src/pybind/mgr/dashboard/frontend/src/app/shared/enum/nvmeof.enum.ts
@@ -1,10 +1,0 @@
-export enum NvmeofSubsystemAuthType {
-  NO_AUTH = 'No authentication',
-  UNIDIRECTIONAL = 'Unidirectional',
-  BIDIRECTIONAL = 'Bi-directional'
-}
-
-export enum NvmeofGatewayNodeMode {
-  SELECTOR = 'selector',
-  DETAILS = 'details'
-}
